Move some objects definitions in `pdf.py` to `generics.py`.
Since these many classes and support functions inherited from `PdfObject` it
made sense to move them there, where object definitions are stored. `pdf.py`
is for another thing altogether.
