+++ pypdf/pdf.py
@@ -2245,7 +2245,7 @@ class PdfFileReader(object):
                     # discussion of the W parameter in PDF spec. table 17.
                     if entrySizes[i] > 0:
                         d = streamData.read(entrySizes[i])
-                        return convertToInt(d, entrySizes[i])
+                        return _convertToInt(d, entrySizes[i])
 
                     # PDF Spec Table 17: A value of zero for an element in the
                     # W array indicates... the default value shall be used
@@ -2544,894 +2544,9 @@ class PdfFileReader(object):
     """
 
 
-def getRectangle(self, name, defaults):
-    retval = self.get(name)
-
-    if isinstance(retval, RectangleObject):
-        return retval
-    if retval is None:
-        for d in defaults:
-            retval = self.get(d)
-            if retval is not None:
-                break
-    if isinstance(retval, IndirectObject):
-        retval = self.pdf.getObject(retval)
-
-    retval = RectangleObject(retval)
-    setRectangle(self, name, retval)
-
-    return retval
-
-
-def setRectangle(self, name, value):
-    if not isinstance(name, NameObject):
-        name = NameObject(name)
-    self[name] = value
-
-
-def deleteRectangle(self, name):
-    del self[name]
-
-
-def createRectangleAccessor(name, fallback):
-    return property(
-        lambda self: getRectangle(self, name, fallback),
-        lambda self, value: setRectangle(self, name, value),
-        lambda self: deleteRectangle(self, name)
-    )
-
-
-class PageObject(DictionaryObject):
-    """
-    This class represents a single page within a PDF file.  Typically this
-    object will be created by accessing the
-    :meth:`getPage()<pypdf.PdfFileReader.getPage>` method of the
-    :class:`PdfFileReader<pypdf.PdfFileReader>` class, but it is
-    also possible to create an empty page with the
-    :meth:`createBlankPage()<PageObject.createBlankPage>` static method.
-
-    :param pdf: PDF file the page belongs to.
-    :param indirectRef: Stores the original indirect reference to
-        this object in its source PDF
-    """
-    def __init__(self, pdf=None, indirectRef=None):
-        DictionaryObject.__init__(self)
-        self.pdf = pdf
-        self.indirectRef = indirectRef
-
-    @staticmethod
-    def createBlankPage(pdf=None, width=None, height=None):
-        """
-        Returns a new blank page.
-        If ``width`` or ``height`` is ``None``, try to get the page size from
-        the last page of *pdf*.
-
-        :param pdf: PDF file the page belongs to
-        :param float width: The width of the new page expressed in default user
-            space units.
-        :param float height: The height of the new page expressed in default
-            user space units.
-        :return: the new blank page:
-        :rtype: :class:`PageObject<PageObject>`
-        :raises PageSizeNotDefinedError: if ``pdf`` is ``None`` or contains
-            no page
-        """
-        page = PageObject(pdf)
-
-        # Creates a new page (cnf. PDF Reference 7.7.3.3)
-        page.__setitem__(NameObject('/Type'), NameObject('/Page'))
-        page.__setitem__(NameObject('/Parent'), NullObject())
-        page.__setitem__(NameObject('/Resources'), DictionaryObject())
-
-        if width is None or height is None:
-            if pdf is not None and pdf.getNumPages() > 0:
-                lastpage = pdf.getPage(pdf.getNumPages() - 1)
-                width = lastpage.mediaBox.getWidth()
-                height = lastpage.mediaBox.getHeight()
-            else:
-                raise PageSizeNotDefinedError()
-        page.__setitem__(
-            NameObject('/MediaBox'), RectangleObject([0, 0, width, height])
-        )
-
-        return page
-
-    def rotateClockwise(self, angle):
-        """
-        Rotates a page clockwise by increments of 90 degrees.
-
-        :param int angle: Angle to rotate the page.  Must be an increment of 90
-            deg.
-        """
-        assert angle % 90 == 0
-        self._rotate(angle)
-        return self
-
-    def rotateCounterClockwise(self, angle):
-        """
-        Rotates a page counter-clockwise by increments of 90 degrees.
-
-        :param int angle: Angle to rotate the page.  Must be an increment
-            of 90 deg.
-        """
-        assert angle % 90 == 0
-        self._rotate(-angle)
-        return self
-
-    def _rotate(self, angle):
-        rotateObj = self.get("/Rotate", 0)
-        currentAngle = rotateObj if isinstance(rotateObj, int) else\
-            rotateObj.getObject()
-        self[NameObject("/Rotate")] = NumberObject(currentAngle + angle)
-
-    @staticmethod
-    def _mergeResources(res1, res2, resource):
-        newRes = DictionaryObject()
-        newRes.update(res1.get(resource, DictionaryObject()).getObject())
-        page2Res = res2.get(resource, DictionaryObject()).getObject()
-        renameRes = {}
-
-        for key in list(page2Res.keys()):
-            if key in newRes and newRes.rawGet(key) != page2Res.rawGet(key):
-                newname = NameObject(key + str(uuid.uuid4()))
-                renameRes[key] = newname
-                newRes[newname] = page2Res[key]
-            elif key not in newRes:
-                newRes[key] = page2Res.rawGet(key)
-
-        return newRes, renameRes
-
-    @staticmethod
-    def _contentStreamRename(stream, rename, pdf):
-        if not rename:
-            return stream
-
-        stream = ContentStream(stream, pdf)
-
-        for operands, operator in stream.operations:
-            for i in range(len(operands)):
-                op = operands[i]
-                if isinstance(op, NameObject):
-                    operands[i] = rename.get(op, op)
-
-        return stream
-
-    @staticmethod
-    def _pushPopGS(contents, pdf):
-        # Adds a graphics state "push" and "pop" to the beginning and end of a
-        # content stream.  This isolates it from changes such as transformation
-        # matrices.
-        stream = ContentStream(contents, pdf)
-        stream.operations.insert(0, [[], "q"])
-        stream.operations.append([[], "Q"])
-
-        return stream
-
-    @staticmethod
-    def _addTransformationMatrix(contents, pdf, ctm):
-        # Adds transformation matrix at the beginning of the given contents
-        # stream.
-        a, b, c, d, e, f = ctm
-        contents = ContentStream(contents, pdf)
-        contents.operations.insert(
-            0, [[FloatObject(a), FloatObject(b), FloatObject(c),
-                 FloatObject(d), FloatObject(e), FloatObject(f)], " cm"]
-        )
-
-        return contents
-
-    def getContents(self):
-        """
-        Accesses the page contents.
-
-        :return: the ``/Contents`` object, or ``None`` if it doesn't exist.
-            ``/Contents`` is optional, as described in PDF Reference  7.7.3.3
-        """
-        if "/Contents" in self:
-            return self["/Contents"].getObject()
-        else:
-            return None
-
-    def mergePage(self, page2):
-        """
-        Merges the content streams of two pages into one.  Resource references
-        (i.e. fonts) are maintained from both pages.  The mediabox/cropbox/etc
-        of this page are not altered.  The parameter page's content stream will
-        be added to the end of this page's content stream, meaning that it will
-        be drawn after, or "on top" of this page.
-
-        :param PageObject page2: The page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        """
-        self._mergePage(page2)
-
-    def _mergePage(self, page2, page2transformation=None, ctm=None,
-                   expand=False):
-        # First we work on merging the resource dictionaries.  This allows us
-        # to find out what symbols in the content streams we might need to
-        # rename.
-        newResources = DictionaryObject()
-        rename = {}
-        originalResources = self["/Resources"].getObject()
-        page2Resources = page2["/Resources"].getObject()
-        newAnnots = ArrayObject()
-
-        for page in (self, page2):
-            if "/Annots" in page:
-                annots = page["/Annots"]
-                if isinstance(annots, ArrayObject):
-                    for ref in annots:
-                        newAnnots.append(ref)
-
-        for res in ("/ExtGState", "/Font", "/XObject", "/ColorSpace",
-                    "/Pattern", "/Shading", "/Properties"):
-            new, newrename = PageObject._mergeResources(
-                originalResources, page2Resources, res
-            )
-            if new:
-                newResources[NameObject(res)] = new
-                rename.update(newrename)
-
-        # Combine /ProcSet sets.
-        newResources[NameObject("/ProcSet")] = ArrayObject(
-            frozenset(
-                originalResources.get("/ProcSet", ArrayObject()).getObject()
-            ).union(
-                frozenset(
-                    page2Resources.get("/ProcSet", ArrayObject()).getObject()
-                )
-            )
-        )
-
-        newContentArray = ArrayObject()
-
-        originalContent = self.getContents()
-
-        if originalContent is not None:
-            newContentArray.append(PageObject._pushPopGS(
-                  originalContent, self.pdf))
-
-        page2Content = page2.getContents()
-
-        if page2Content is not None:
-            if page2transformation is not None:
-                page2Content = page2transformation(page2Content)
-            page2Content = PageObject._contentStreamRename(
-                page2Content, rename, self.pdf)
-            page2Content = PageObject._pushPopGS(page2Content, self.pdf)
-            newContentArray.append(page2Content)
-
-        # If expanding the page to fit a new page, calculate the new media box size
-        if expand:
-            corners1 = [self.mediaBox.getLowerLeft_x().asNumeric(),
-                        self.mediaBox.getLowerLeft_y().asNumeric(),
-                        self.mediaBox.getUpperRight_x().asNumeric(),
-                        self.mediaBox.getUpperRight_y().asNumeric()]
-            corners2 = [page2.mediaBox.getLowerLeft_x().asNumeric(),
-                        page2.mediaBox.getLowerLeft_y().asNumeric(),
-                        page2.mediaBox.getUpperLeft_x().asNumeric(),
-                        page2.mediaBox.getUpperLeft_y().asNumeric(),
-                        page2.mediaBox.getUpperRight_x().asNumeric(),
-                        page2.mediaBox.getUpperRight_y().asNumeric(),
-                        page2.mediaBox.getLowerRight_x().asNumeric(),
-                        page2.mediaBox.getLowerRight_y().asNumeric()]
-            if ctm is not None:
-                ctm = [float(x) for x in ctm]
-                new_x = [ctm[0]*corners2[i] + ctm[2]*corners2[i+1] + ctm[4]
-                         for i in range(0, 8, 2)]
-                new_y = [ctm[1]*corners2[i] + ctm[3]*corners2[i+1] + ctm[5]
-                         for i in range(0, 8, 2)]
-            else:
-                new_x = corners2[0:8:2]
-                new_y = corners2[1:8:2]
-
-            lowerleft = [min(new_x), min(new_y)]
-            upperright = [max(new_x), max(new_y)]
-            lowerleft = [min(corners1[0], lowerleft[0]), min(corners1[1],
-                                                             lowerleft[1])]
-            upperright = [max(corners1[2], upperright[0]), max(corners1[3],
-                                                               upperright[1])]
-
-            self.mediaBox.setLowerLeft(lowerleft)
-            self.mediaBox.setUpperRight(upperright)
-
-        self[NameObject('/Contents')] = ContentStream(newContentArray,
-                                                      self.pdf)
-        self[NameObject('/Resources')] = newResources
-        self[NameObject('/Annots')] = newAnnots
-
-    def mergeTransformedPage(self, page2, ctm, expand=False):
-        """
-        This is similar to mergePage, but a transformation matrix is
-        applied to the merged stream.
-
-        :param PageObject page2: The page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param tuple ctm: a 6-element tuple containing the operands of the
-            transformation matrix
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        self._mergePage(
-            page2, lambda page2Content: PageObject._addTransformationMatrix(
-                page2Content, page2.pdf, ctm), ctm, expand
-        )
-
-    def mergeScaledPage(self, page2, scale, expand=False):
-        """
-        This is similar to mergePage, but the stream to be merged is scaled
-        by appling a transformation matrix.
-
-        :param PageObject page2: The page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float scale: The scaling factor
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        # CTM to scale : [ sx 0 0 sy 0 0 ]
-        return self.mergeTransformedPage(
-            page2, (scale, 0, 0, scale, 0, 0), expand
-        )
-
-    def mergeRotatedPage(self, page2, rotation, expand=False):
-        """
-        This is similar to mergePage, but the stream to be merged is rotated
-        by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float rotation: The angle of the rotation, in degrees
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        rotation = math.radians(rotation)
-
-        return self.mergeTransformedPage(
-            page2, (math.cos(rotation), math.sin(rotation),
-                    -math.sin(rotation), math.cos(rotation), 0, 0), expand
-        )
-
-    def mergeTranslatedPage(self, page2, tx, ty, expand=False):
-        """
-        This is similar to ``mergePage``, but the stream to be merged is
-        translated by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float tx: The translation on X axis.
-        :param float ty: The translation on Y axis.
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        return self.mergeTransformedPage(page2, (1,  0, 0,  1, tx, ty), expand)
-
-    def mergeRotatedTranslatedPage(
-            self, page2, rotation, tx, ty, expand=False
-    ):
-        """
-        This is similar to mergePage, but the stream to be merged is rotated
-        and translated by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float tx: The translation on X axis.
-        :param float ty: The translation on Y axis.
-        :param float rotation: The angle of the rotation, in degrees.
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-
-        translation = [[1, 0, 0], [0, 1, 0], [-tx, -ty, 1]]
-        rotation = math.radians(rotation)
-        rotating = [
-            [math.cos(rotation), math.sin(rotation), 0],
-            [-math.sin(rotation), math.cos(rotation), 0],
-            [0, 0, 1]
-        ]
-        rtranslation = [
-            [1, 0, 0], [0, 1, 0], [tx, ty, 1]
-        ]
-        ctm = matrixMultiply(translation, rotating)
-        ctm = matrixMultiply(ctm, rtranslation)
-
-        return self.mergeTransformedPage(
-            page2,
-            (ctm[0][0], ctm[0][1], ctm[1][0], ctm[1][1], ctm[2][0], ctm[2][1]),
-            expand
-        )
-
-    def mergeRotatedScaledPage(self, page2, rotation, scale, expand=False):
-        """
-        This is similar to mergePage, but the stream to be merged is rotated
-        and scaled by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float rotation: The angle of the rotation, in degrees.
-        :param float scale: The scaling factor.
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        rotation = math.radians(rotation)
-        rotating = [
-            [math.cos(rotation), math.sin(rotation), 0],
-            [-math.sin(rotation), math.cos(rotation), 0],
-            [0, 0, 1]
-        ]
-        scaling = [
-            [scale, 0, 0], [0, scale, 0], [0, 0, 1]
-        ]
-        ctm = matrixMultiply(rotating, scaling)
-
-        return self.mergeTransformedPage(
-            page2,
-            (ctm[0][0], ctm[0][1], ctm[1][0], ctm[1][1], ctm[2][0], ctm[2][1]),
-            expand
-        )
-
-    def mergeScaledTranslatedPage(self, page2, scale, tx, ty, expand=False):
-        """
-        This is similar to mergePage, but the stream to be merged is translated
-        and scaled by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float scale: The scaling factor.
-        :param float tx: The translation on X axis.
-        :param float ty: The translation on Y axis.
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-
-        translation = [
-            [1, 0, 0], [0, 1, 0], [tx, ty, 1]
-        ]
-        scaling = [
-            [scale, 0, 0], [0, scale, 0], [0, 0, 1]
-        ]
-        ctm = matrixMultiply(scaling, translation)
-
-        return self.mergeTransformedPage(
-            page2,
-            (ctm[0][0], ctm[0][1], ctm[1][0], ctm[1][1], ctm[2][0], ctm[2][1]),
-            expand
-        )
-
-    def mergeRotatedScaledTranslatedPage(
-            self, page2, rotation, scale, tx, ty, expand=False
-    ):
-        """
-        This is similar to mergePage, but the stream to be merged is
-        translated, rotated and scaled by appling a transformation matrix.
-
-        :param PageObject page2: the page to be merged into this one. Should be
-            an instance of :class:`PageObject<PageObject>`.
-        :param float tx: The translation on X axis.
-        :param float ty: The translation on Y axis.
-        :param float rotation: The angle of the rotation, in degrees.
-        :param float scale: The scaling factor.
-        :param bool expand: Whether the page should be expanded to fit the
-            dimensions of the page to be merged.
-        """
-        translation = [
-            [1, 0, 0], [0, 1, 0], [tx, ty, 1]
-        ]
-        rotation = math.radians(rotation)
-        rotating = [
-            [math.cos(rotation), math.sin(rotation), 0],
-            [-math.sin(rotation), math.cos(rotation), 0],
-            [0, 0, 1]
-        ]
-        scaling = [
-            [scale, 0, 0], [0, scale, 0], [0, 0, 1]
-        ]
-        ctm = matrixMultiply(rotating, scaling)
-        ctm = matrixMultiply(ctm, translation)
-
-        return self.mergeTransformedPage(
-            page2,
-            (ctm[0][0], ctm[0][1], ctm[1][0], ctm[1][1], ctm[2][0], ctm[2][1]),
-            expand
-        )
-
-    def addTransformation(self, ctm):
-        """
-        Applies a transformation matrix to the page.
-
-        :param tuple ctm: A 6-element tuple containing the operands of the
-            transformation matrix.
-        """
-        originalContent = self.getContents()
-
-        if originalContent is not None:
-            newContent = PageObject._addTransformationMatrix(
-                originalContent, self.pdf, ctm)
-            newContent = PageObject._pushPopGS(newContent, self.pdf)
-            self[NameObject('/Contents')] = newContent
-
-    def scale(self, sx, sy):
-        """
-        Scales a page by the given factors by appling a transformation
-        matrix to its content and updating the page size.
-
-        :param float sx: The scaling factor on horizontal axis.
-        :param float sy: The scaling factor on vertical axis.
-        """
-        self.addTransformation((sx, 0, 0,  sy, 0,  0))
-        self.mediaBox = RectangleObject([
-            float(self.mediaBox.getLowerLeft_x()) * sx,
-            float(self.mediaBox.getLowerLeft_y()) * sy,
-            float(self.mediaBox.getUpperRight_x()) * sx,
-            float(self.mediaBox.getUpperRight_y()) * sy]
-        )
-
-        if "/VP" in self:
-            viewport = self["/VP"]
-
-            if isinstance(viewport, ArrayObject):
-                bbox = viewport[0]["/BBox"]
-            else:
-                bbox = viewport["/BBox"]
-
-            scaled_bbox = RectangleObject([
-                float(bbox[0]) * sx, float(bbox[1]) * sy, float(bbox[2]) * sx,
-                float(bbox[3]) * sy]
-            )
-
-            if isinstance(viewport, ArrayObject):
-                self[NameObject("/VP")][NumberObject(0)][NameObject("/BBox")]\
-                    = scaled_bbox
-            else:
-                self[NameObject("/VP")][NameObject("/BBox")] = scaled_bbox
-
-    def scaleBy(self, factor):
-        """
-        Scales a page by the given factor by appling a transformation
-        matrix to its content and updating the page size.
-
-        :param float factor: The scaling factor (for both X and Y axis).
-        """
-        self.scale(factor, factor)
-
-    def scaleTo(self, width, height):
-        """
-        Scales a page to the specified dimentions by appling a
-        transformation matrix to its content and updating the page size.
-
-        :param float width: The new width.
-        :param float height: The new heigth.
-        """
-        sx = width / float(self.mediaBox.getUpperRight_x() -
-                           self.mediaBox.getLowerLeft_x())
-        sy = height / float(self.mediaBox.getUpperRight_y() -
-                            self.mediaBox.getLowerLeft_y())
-        self.scale(sx, sy)
-
-    def compressContentStreams(self):
-        """
-        Compresses the size of this page by joining all content streams and
-        applying a FlateDecode filter.
-
-        However, it is possible that this function will perform no action if
-        content stream compression becomes "automatic" for some reason.
-        """
-        content = self.getContents()
-
-        if content is not None:
-            if not isinstance(content, ContentStream):
-                content = ContentStream(content, self.pdf)
-            self[NameObject("/Contents")] = content.flateEncode()
-
-    def extractText(self):
-        """
-        Locate all text drawing commands, in the order they are provided in the
-        content stream, and extract the text.  This works well for some PDF
-        files, but poorly for others, depending on the generator used.  This
-        will be refined in the future.  Do not rely on the order of text coming
-        out of this function, as it will change if this function is made more
-        sophisticated.
-
-        :return: a unicode string object.
-        """
-        text = pypdfUnicode("")
-        content = self["/Contents"].getObject()
-
-        if not isinstance(content, ContentStream):
-            content = ContentStream(content, self.pdf)
-
-        # Note: we check all strings are TextStringObjects.  ByteStringObjects
-        # are strings where the byte->string encoding was unknown, so adding
-        # them to the text here would be gibberish.
-        for operands, operator in content.operations:
-            if operator == b_("Tj"):
-                _text = operands[0]
-
-                if isinstance(_text, TextStringObject):
-                    text += _text
-                    text += "\n"
-            elif operator == b_("T*"):
-                text += "\n"
-            elif operator == b_("'"):
-                text += "\n"
-                _text = operands[0]
-                if isinstance(_text, TextStringObject):
-                    text += operands[0]
-            elif operator == b_('"'):
-                _text = operands[2]
-                if isinstance(_text, TextStringObject):
-                    text += "\n"
-                    text += _text
-            elif operator == b_("TJ"):
-                for i in operands[0]:
-                    if isinstance(i, TextStringObject):
-                        text += i
-                text += "\n"
-
-        return text
-
-    mediaBox = createRectangleAccessor("/MediaBox", ())
-    """
-    A :class:`RectangleObject<pypdf.generic.RectangleObject>`, expressed in
-    default user space units, defining the boundaries of the physical medium on
-    which the page is intended to be displayed or printed.
-    """
-
-    cropBox = createRectangleAccessor("/CropBox", ("/MediaBox",))
-    """
-    A :class:`RectangleObject<pypdf.generic.RectangleObject>`, expressed in
-    default user space units, defining the visible region of default user
-    space.  When the page is displayed or printed, its contents are to be
-    clipped (cropped) to this rectangle and then imposed on the output medium
-    in some implementation-defined manner.  Default value: same as
-    :attr:`mediaBox<mediaBox>`.
-    """
-
-    bleedBox = createRectangleAccessor("/BleedBox", ("/CropBox", "/MediaBox"))
-    """
-    A :class:`RectangleObject<pypdf.generic.RectangleObject>`, expressed in
-    default user space units, defining the region to which the contents of the
-    page should be clipped when output in a production enviroment.
-    """
-
-    trimBox = createRectangleAccessor("/TrimBox", ("/CropBox", "/MediaBox"))
-    """
-    A :class:`RectangleObject<pypdf.generic.RectangleObject>`, expressed in
-    default user space units, defining the intended dimensions of the finished
-    page after trimming.
-    """
-
-    artBox = createRectangleAccessor("/ArtBox", ("/CropBox", "/MediaBox"))
-    """
-    A :class:`RectangleObject<pypdf.generic.RectangleObject>`, expressed in
-    default user space units, defining the extent of the page's meaningful
-    content as intended by the page's creator.
-    """
-
-
-class ContentStream(DecodedStreamObject):
-    def __init__(self, stream, pdf):
-        # TO-DO Why is the call to super().__init__() missing here? Shouldn't
-        # it be added?
-        self.pdf = pdf
-        self.operations = []
-        # stream may be a StreamObject or an ArrayObject containing
-        # multiple StreamObjects to be cat'd together.
-        stream = stream.getObject()
-
-        if isinstance(stream, ArrayObject):
-            data = b_("")
-            for s in stream:
-                data += b_(s.getObject().getData())
-            stream = BytesIO(b_(data))
-        else:
-            stream = BytesIO(b_(stream.getData()))
-
-        self.__parseContentStream(stream)
-
-    def __parseContentStream(self, stream):
-        stream.seek(0, 0)
-        operands = []
-
-        while True:
-            peek = readNonWhitespace(stream)
-
-            if peek == b_('') or pypdfOrd(peek) == 0:
-                break
-
-            stream.seek(-1, 1)
-            if peek.isalpha() or peek == b_("'") or peek == b_('"'):
-                operator = readUntilRegex(
-                    stream, NameObject.delimiterPattern, True
-                )
-                if operator == b_("BI"):
-                    # Begin inline image - a completely different parsing
-                    # mechanism is required
-                    assert operands == []
-                    ii = self._readInlineImage(stream)
-                    self.operations.append((ii, b_("INLINE IMAGE")))
-                else:
-                    self.operations.append((operands, operator))
-                    operands = []
-            elif peek == b_('%'):
-                # If we encounter a comment in the content stream, we have to
-                # handle it here.  Typically, readObject will handle
-                # encountering a comment -- but readObject assumes that
-                # following the comment must be the object we're trying to
-                # read.  In this case, it could be an operator instead.
-                while peek not in (b_('\r'), b_('\n')):
-                    peek = stream.read(1)
-            else:
-                operands.append(readObject(stream, None))
-
-    def _readInlineImage(self, stream):
-        # Begin reading just after the "BI" - begin image
-        # First read the dictionary of settings.
-        settings = DictionaryObject()
-
-        while True:
-            tok = readNonWhitespace(stream)
-            stream.seek(-1, 1)
-
-            if tok == b_("I"):
-                # "ID" - begin of image data
-                break
-
-            key = readObject(stream, self.pdf)
-            tok = readNonWhitespace(stream)
-            stream.seek(-1, 1)
-            value = readObject(stream, self.pdf)
-            settings[key] = value
-
-        # Left at beginning of ID
-        tmp = stream.read(3)
-        assert tmp[:2] == b_("ID")
-        data = b_("")
-
-        while True:
-            # Read the inline image, while checking for EI (End Image) operator
-            tok = stream.read(1)
-
-            if tok == b_("E"):
-                # Check for End Image
-                tok2 = stream.read(1)
-                if tok2 == b_("I"):
-                    # Data can contain EI, so check for the Q operator.
-                    tok3 = stream.read(1)
-                    info = tok + tok2
-                    # We need to find whitespace between EI and Q.
-                    has_q_whitespace = False
-
-                    while tok3 in WHITESPACES:
-                        has_q_whitespace = True
-                        info += tok3
-                        tok3 = stream.read(1)
-                    if tok3 == b_("Q") and has_q_whitespace:
-                        stream.seek(-1, 1)
-                        break
-                    else:
-                        stream.seek(-1, 1)
-                        data += info
-                else:
-                    stream.seek(-1, 1)
-                    data += tok
-            else:
-                data += tok
-
-        return {"settings": settings, "data": data}
-
-    def _getData(self):
-        newdata = BytesIO()
-
-        for operands, operator in self.operations:
-            if operator == b_("INLINE IMAGE"):
-                newdata.write(b_("BI"))
-                dicttext = BytesIO()
-                operands["settings"].writeToStream(dicttext, None)
-                newdata.write(dicttext.getvalue()[2:-2])
-                newdata.write(b_("ID "))
-                newdata.write(operands["data"])
-                newdata.write(b_("EI"))
-            else:
-                for op in operands:
-                    op.writeToStream(newdata, None)
-                    newdata.write(b_(" "))
-
-                newdata.write(b_(operator))
-
-            newdata.write(b_("\n"))
-
-        return newdata.getvalue()
-
-    def _setData(self, value):
-        self.__parseContentStream(BytesIO(b_(value)))
-
-    _data = property(_getData, _setData)
-
-
-class DocumentInformation(DictionaryObject):
-    """
-    A class representing the basic document metadata provided in a PDF File.
-    This class is accessible through
-    :meth:`getDocumentInfo()<pypdf.PdfFileReader.getDocumentInfo()>`
-
-    All text properties of the document metadata have
-    *two* properties, eg. author and author_raw. The non-raw property will
-    always return a ``TextStringObject``, making it ideal for a case where
-    the metadata is being displayed. The raw property can sometimes return
-    a ``ByteStringObject``, if PyPDF was unable to decode the string's
-    text encoding; this requires additional safety in the caller and
-    therefore is not as commonly accessed.
-    """
-
-    def __init__(self):
-        DictionaryObject.__init__(self)
-
-    def getText(self, key):
-        retval = self.get(key, None)
-
-        if isinstance(retval, TextStringObject):
-            return retval
-
-        return None
-
-    title = property(lambda self: self.getText("/Title"))
-    """
-    Read-only property accessing the document's **title**.
-    Returns a unicode string (``TextStringObject``) or ``None``
-    if the title is not specified.
-    """
-
-    title_raw = property(lambda self: self.get("/Title"))
-    """The "raw" version of title; can return a ``ByteStringObject``."""
-
-    author = property(lambda self: self.getText("/Author"))
-    """
-    Read-only property accessing the document's **author**.
-    Returns a unicode string (``TextStringObject``) or ``None``
-    if the author is not specified.
-    """
-
-    author_raw = property(lambda self: self.get("/Author"))
-    """The "raw" version of author; can return a ``ByteStringObject``."""
-
-    subject = property(lambda self: self.getText("/Subject"))
-    """
-    Read-only property accessing the document's **subject**.
-    Returns a unicode string (``TextStringObject``) or ``None``
-    if the subject is not specified.
-    """
-
-    subject_raw = property(lambda self: self.get("/Subject"))
-    """The "raw" version of subject; can return a ``ByteStringObject``."""
-
-    creator = property(lambda self: self.getText("/Creator"))
-    """
-    Read-only property accessing the document's **creator**. If the
-    document was converted to PDF from another format, this is the name of the
-    application (e.g. OpenOffice) that created the original document from
-    which it was converted. Returns a unicode string (``TextStringObject``)
-    or ``None`` if the creator is not specified.
-    """
-
-    creator_raw = property(lambda self: self.get("/Creator"))
-    """The "raw" version of creator; can return a ``ByteStringObject``."""
-
-    producer = property(lambda self: self.getText("/Producer"))
-    """
-    Read-only property accessing the document's **producer**.
-    If the document was converted to PDF from another format, this is
-    the name of the application (for example, OSX Quartz) that converted
-    it to PDF. Returns a unicode string (``TextStringObject``)
-    or ``None`` if the producer is not specified.
-    """
-
-    producer_raw = property(lambda self: self.get("/Producer"))
-    """The "raw" version of producer; can return a ``ByteStringObject``."""
-
-
-def convertToInt(d, size):
+def _convertToInt(d, size):
     if size > 8:
-        raise PdfReadError("Invalid size in convertToInt")
+        raise PdfReadError("Invalid size in _convertToInt")
 
     d = b_("\x00\x00\x00\x00\x00\x00\x00\x00") + b_(d)
     d = d[-8:]
